+++ crate/clients/ckms/src/lib.rs
@@ -2,6 +2,7 @@ pub mod actions;
 pub mod commands;
 pub mod config;
 pub mod error;
+pub mod headers_config;
 pub mod proxy_config;
 
 pub use commands::{Cli, CliCommands, ckms_main};